feat: Add Styled Component library to create the design system theme
Styled-components is a library for React and React Native that allows
you to use styles in component format in your application.
It allows you to write JavaScript and CSS together, has Sass syntax for
grouping classes and also makes it easy to create light and dark themes
using global styles and theme provider.